Move assembly stripper to a separate assembly
Build it with nullability checking disabled.

Build the main MonoTargetsTasks assembly with nullability enabled
