+++ src/tasks/MonoTargetsTasks/MonoTargetsTasks.csproj
@@ -4,11 +4,10 @@
     <OutputType>Library</OutputType>
     <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
     <EnableDefaultCompileItems>false</EnableDefaultCompileItems>
-    <!-- FIXME: CilStrip.Sources doesn't play nicely with nullability - build it to a separate .dll -->
-    <Nullable>disable</Nullable>
-    <NoWarn>$(NoWarn),CA1050,CS0618,CS0649,CS8604,CS8602,CS8632,SYSLIB0003</NoWarn>
+    <Nullable>enable</Nullable>
+    <NoWarn>$(NoWarn),CA1050,CS0618,CS0649,CS8632,SYSLIB0003</NoWarn>
     <!-- Ignore nullable warnings on net4* -->
-    <!-- // NoWarn Condition="$(TargetFramework.StartsWith('net4'))">$(NoWarn),CS8604,CS8602</NoWarn -->
+    <NoWarn Condition="$(TargetFramework.StartsWith('net4'))">$(NoWarn),CS8604,CS8602</NoWarn>
     <RunAnalyzers>false</RunAnalyzers>
   </PropertyGroup>
   <ItemGroup>
@@ -28,6 +27,10 @@
     <!-- These versions should not be older than what Visual Studio MSBuild uses -->
     <PackageReference Include="System.Threading.Tasks.Extensions" Version="4.5.4" PrivateAssets="all" />
     <PackageReference Include="System.Text.Json" Version="5.0.0" PrivateAssets="all" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="ILStrip\AssemblyStripper\AssemblyStripper.csproj" />
   </ItemGroup>
 
   <ItemGroup>
